+++ saffier/db/fields.py
@@ -3,6 +3,7 @@ import re
 import typing
 from math import isfinite
 
+
 from saffier.core import formats
 from saffier.core.base import ValidationResult
 from saffier.core.datastructures import ArbitraryHashableBaseModel