+++ docs/release-notes.md
@@ -1,5 +1,15 @@
 # Release Notes
 
+## 0.2.0
+
+### Added
+
+- New [Index](./models.md#indexes) object allowing the creation of internal SQLAlchemy indexes.
+  
+### Changed
+
+- Updated metaclass to validate the fields being added to `indexes`.
+
 ## 0.1.0
 
 This is the initial release of Saffier.